методы addFriend,deleteFriend,getFriends,getCommonFriends вынесены из UserRepository и InDbUserRepository в UserFriendRepository
все тесты postman прошли.
