+++ src/main/java/ru/yandex/practicum/filmorate/repository/UserRepository.java
@@ -4,7 +4,6 @@ import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 public interface UserRepository {
 
@@ -16,12 +15,12 @@ public interface UserRepository {
 
     User update(User user);
 
-    void addFriend(User user, User friend, boolean isConfirmed);
+//    void addFriend(User user, User friend, boolean isConfirmed);
 
-    void deleteFriend(User user, User friend);
+//    void deleteFriend(User user, User friend);
 
-    Set<User> getFriends(User user);
+//    Set<User> getFriends(User user);
 
-    List<User> getCommonFriends(long userId, long otherId);
+//    List<User> getCommonFriends(long userId, long otherId);
 
 }
