+++ src/main/java/ru/yandex/practicum/filmorate/service/UserServiceImpl.java
@@ -4,16 +4,21 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.model.UserFriend;
+import ru.yandex.practicum.filmorate.repository.UserFriendRepository;
 import ru.yandex.practicum.filmorate.repository.UserRepository;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final UserFriendRepository userFriendRepository;
 
 
     public List<User> getAll() {
@@ -47,7 +52,7 @@ public class UserServiceImpl implements UserService {
         final User friend = userRepository.get(friendId)
                 .orElseThrow(() -> new ValidationException("Пользователь c id: " + friendId + " не найден"));
 
-        userRepository.addFriend(user, friend, isConfirmed);
+        userFriendRepository.addFriend(user, friend, isConfirmed);
     }
 
     public void deleteFriend(long userId, long friendId) {
@@ -57,7 +62,7 @@ public class UserServiceImpl implements UserService {
         User friend = userRepository.get(friendId)
                 .orElseThrow(() -> new ValidationException("Пользователь c id: " + friendId + " не найден"));
 
-        userRepository.deleteFriend(user, friend);
+        userFriendRepository.deleteFriend(user, friend);
     }
 
     public Set<User> getFriends(long userId) {
@@ -65,7 +70,13 @@ public class UserServiceImpl implements UserService {
         final User user = userRepository.get(userId)
                 .orElseThrow(() -> new ValidationException("Пользователь c id: " + userId + " не найден"));
 
-        return userRepository.getFriends(user);
+        Set<Long> uf = userFriendRepository.getFriends(user).stream()
+                .map(UserFriend::getFriendId)
+                .collect(Collectors.toSet());
+
+        return userRepository.getAll().stream()
+                .filter(u -> uf.contains(u.getId()))
+                .collect(Collectors.toSet());
     }
 
     public List<User> getCommonFriends(long userId, long otherId) {
@@ -76,7 +87,17 @@ public class UserServiceImpl implements UserService {
         User other = userRepository.get(otherId)
                 .orElseThrow(() -> new ValidationException("Пользователь c id: " + otherId + " не найден"));
 
-        return userRepository.getCommonFriends(user.getId(), other.getId());
+        Set<Long> uf = userFriendRepository.getCommonFriends(user.getId(), other.getId()).stream()
+                .map(UserFriend::getFriendId)
+                .collect(Collectors.groupingBy(friendId -> friendId, Collectors.counting())) // Группируем по friendId и считаем
+                .entrySet().stream() // Преобразуем в поток пар (friendId, count)
+                .filter(entry -> entry.getValue() > 1) // Оставляем только те, где count > 1
+                .map(Map.Entry::getKey) // Извлекаем friendId
+                .collect(Collectors.toSet()); // Собираем в Set
+
+        return userRepository.getAll().stream()
+                .filter(u -> uf.contains(u.getId()))
+                .toList();
     }
 
     private void checkUserName(User user) {
